I realize this may be rather 'dumb' - adding qrencode's config.h to svn, since it's likely specific to the CentOS variant I'm running. However, it's easy enough to regen locally if the config.h in subversion doesn't match your local machine. Just use the 'configure' script in the qrencode-... folder to regen config.h if you have qrencode-specific compile problems.

+++ trunk/src/glvidtex/GLRectDrawable.cpp
@@ -49,7 +49,7 @@ void GLRectDrawable::renderImage()
 		{
 			painter.setPen(QPen(m_borderColor,m_borderWidth));
 			painter.setBrush(m_fillColor);
-			int v = m_borderWidth / 2;
+			int v = (int)m_borderWidth / 2;
 			target = target.adjusted(v,v,-v,-v);
 			painter.drawRect(target);
 		}
